changed a unit test to an integration test
